pyproject.toml
[project]
name = "rag-chat"
version = "0.1.0"
description = "RAG Chat with Ollama, Neo4j and Langchain"
authors = [
    { name = "Damian Lebiedź" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "sse-starlette (>=2.3.6,<2.4.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-core (>=0.3.68,<0.4.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "langchain-neo4j (>=0.4.0,<0.5.0)",
    "neo4j (>=5.28.1,<5.29.0)",
    "python-dotenv (>=1.1.1,<1.2.0)",
    "pydantic (>=2.11.7,<2.12.0)",
    "streamlit (==1.32.1)"
]


[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
